Allow regeneration of DB connections if we have given out too many handles to the same connection.
Thanks to <victoriggy@gmail.com> for the idea and the original
implementation. Refactored extensively and max_requests renamed to
max_handles to better represent the functionality. <robbat2>

Further refactored by dormando. Now is a measure of how many times
a handle may be *validated*. Validation is the known safe point for
when a database handle may be reset.

git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1348 f67b2e87-0811-0410-a7e0-dd94e48410d6