don't chew CPU on clusters with many devices.
weighted_list gets real slow when you pass it a lot of devices.
Ideally it should be an iterator, which we'd only end up hitting
a few times. It'd still slow down with many devs but not as badly.

However I want to rework the algorithm to scale against free space
better (probably just throw a log() in there), so lets rework that
later. This change focues writes on the top 20 emptiest devices,
which tends to be what people expect it to to do anyway.

git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1412 f67b2e87-0811-0410-a7e0-dd94e48410d6