remove socket caching for HTTP DELETE operations
HTTPFile's delete routine was reusing the socket cache from the 'sizes'
function, which was setting sockets to nonblock. The delete code is both
unable to fully parse HTTP responses and also relies on the socket being in
blocking mode.

For now, lets stop caching sockets used for delete, and later replace that
whole chunk of code with LWP.

Thanks to Kiall Mac Innes for the bug report and help troubleshooting.

git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1462 f67b2e87-0811-0410-a7e0-dd94e48410d6