Remove forced loading of Store/MySQL.pm
So that line was annoying Postgres/etc users since it forces you to have
DBD::mysql. I removed it, but left a note saying that we have to ensure the DB
driver gets preloaded before daemonization.

It actually does already: ->check_database happens right before the daemonize,
which initializes a store, which kicks off the appropriate `use`. So all we
had to do all this time was remove the stupid line, sorry.