Allow default class to be overridden again.
Also makes its support slightly more DWIM. If you ask to modify the default
class, it will be autoadded to the database if it hasn't yet. Add works as it
should.

Delete *doesn't work*. Once you add a default class to the database, you're
stuck with it. You can continue to change it. We could "delete" it and walk it
back to an original form, but it seems more correct to tell people to not do
that.