Issue 36: Further insulated from autoinc resets
If MySQL is restarted and tempfile loses its autoinc counter, normally
mogilefs will detect that a new fid insertion is a duplicate and resets it to
MAX(fid) FROM file. However this doesn't work as well when your low fids may
be deleted.

So now, trackers check MAX(fid) on start and refuse to go below that. Needs
further work but this will go a long way.