queryworker: reimplement "clear_caches" command
Technically, it "refreshes" the cache and usage information,
but the effect is the same with the new monitor.

By swapping do_monitor_round for clear_caches in the tests,
tests run about 30s faster on my machine.  Additionally,
all changes to devices/hosts/domains/classes should refresh
the cache on a given worker.