don't spew errors if the master goes down
complain if the server had gone away after a ping test. workers will also
crash if the server dies in the middle of its routine, but not continue to
spam.

We were also uselessly selecting slave info from the master and slave
constantly :( Both of those things are probably also jobs for the monitor
worker.