enable TCP keepalives for accepted sockets
Some clients may remain connected and idle for hours/days at a
time.  If their connections drop due to catastrophic failure and
the TCP implementation is unable to notify the server, the
server could eventually run out of file descriptors and/or
memory for new clients.

This change allows castastrophic network failures to be detected
after roughly 2 hours on most setups.  Sysadmins are free to
tweak the TCP keepalive knobs in sysctl or similar.