implement "case_sensitive_list_keys" server setting
Enabling this boolean will make the "after" and "prefix" params
of "list_keys" behave case-sensitively.

If this setting is /not/ enabled, clients will hit
after_mismatch errors when iterating through keys if they are
using an uppercase "prefix" argument and a subsequent list_keys
is called with an "after" that only matches case-insensitively.

If unset, this defaults to false (0) to match historical (buggy)
behavior.  Historical behavior is preserved (even if broken) as
users with small namespaces may rely on case-insensitive
matching.

Postgres users are not affected by this change, as the LIKE
operator in Postgres is always case-sensitive.

This change is tested on all three databases: Postgres, MySQL,
and SQLite.