Add connect timeouts
This should work with the primarily supported databases.

Mainly this is for slave setups, so slaves which are broken but not
immediately RST'ing or error'ing get backed off from.

You'll still get some watchdog hits when a slave first dies, but it'll handle
it properly from there with minimal delays.