New variable that disables devcount updates
Devcount field is only useful for statistics and once you have enough
files it's not that useful even for that. Downing a device generates
less I/O load on the DB when devcount updates are disabled, there aren't
any queries against the file table.

Correct devcount can be restored by disabling the flag and running fsck.