Back off if master DB is down
If monitor worker figures master DB is down, send a flag telling workers to
avoid attempting to connect to the master. If a worker is already connected
and it functions, it will ignore the flag. If a worker is starting up or
reconnecting to the master, it will avoid the attempt.

After this commit a timed-out master DB will finally not cause the tracker to
fail, and reads will continue to work.

Takes almost a minute to notice the master is gone though. some room for
improvement there.