make fsck_checksum == off honored in more places
if fsck_checksum was set to off, it would ignore the checksums deep in the
code, but would still attempt to "fix" the fids each time, which runs far more
code and UPDATE's each fid's devcount even if you tell it not to.

now it does what it should. however FSCK with checksums enabled will still
UPDATE devcount on each check.