fsck: log bad count correctly instead of policy violation
A BCNT error is more descriptive than a generic POVI entry
and more accurately reflects the change made to an FID entry.

This also removes a dependency from /using/ the devcount column
and simplifies the code.  The devcount column remains invaluable
and informative to users, but MogileFS should not trust it for
making decisions when it has access to the file_on table.