worker/query: Add optional callid parameter
Queries are computed in parallel and therefore replies are not in
the right order. This patch adds an optional callid parameter so
a caller can match the replies back to the queries. ERR lines
return the callid as 3rd parameter. If the callid parameter is
missing then the Protocol is the same as before.

Example:
GET_PATHS callid=1
ERR no_domain No+domain+provided 1
GET_DOMAINS callid=1
OK domains=0&callid=1