Delete memcache data when we replicate fids
Because memcache TTL is now user configurable, data in memcached might
be valid for a long time, and as such invalid paths might be returned.

It would be possible to populate memcache, instead of just removing. But
it might be wasteful when a device is marked as dead, those replicated
fids might not need to be in memcached.

There is still one TODO left. If someone modifies mindevcount and runs
FSCK, then the mappings might become incorrect, but I reasoned that it
would be rather rare.