fsck: cleanup and reduce unnecessary devcount updates
fix_fid(): we no longer rely blindly update devcount on every
call.  This is important because we call fix_fid() on checksum
checks regardless, and devcount updates entail unnecessary
updates to the `file' table.

While we're at it, consolidate the places where we check the
skip_devcount flag and log bad devcounts.