Device->observed_* all respects observed host state
The observed unreachable state of the host should always supercede the
observed state of the device.  This is already the case with
observed_writeable, but not with observed_readable nor
observed_unreachable.

The monitor worker does not (and should not, to save bandwidth) update
states of all devices when a host goes down.