reaper: better handling of lock failures
The delay backoff should only occur if we got a successful lock.
Backing off the delay on lock failure can result in the
delay becoming undef and (incorrectly) making the reaper give up
on a device.

Fortunately, lock failures with the extremely long (60s) lock
timeout is unlikely to be a problem in practice.