device: reuse HTTP connections for MKCOL
This can reduce latency for folks still stuck with MKCOL.
This creates no new sockets for replicate and monitor in
all cases, as connections to the HTTP DAV server are already
used in those workers.

This only adds new persistent connections to the queryworker if
GET-only HTTP ports are configured (queryworker already may call
HTTPFile->size).