improve handling of classes which change hash algorithm
In replicate, we validate via existing FID checksum regardless
of class.  This failed when the class.hashtype was altered after
uploading but before replication.

The following sequence of events caused replication to fail:

  1. class.hashtype = MD5
  2. FID created and stores MD5:...
  3. FID enqueued for replication
  4. class.hashtype = NONE
  5. FID begins replicating

Replication (Step 5) failed since the existing MD5 digest is
trying to validate against a (now) non-existent class digest.
Since we stored the checksum in the database anyways, calculate
and validate anyways as an admin could've only wanted to
alter a class temporarily, not permanently.

An admin may also decide to switch checksum algorithms.
Fsck now logs hash algorithm mismatches as "BALG" and
emits a descriptive message to syslog