replicate: use persistent connection from pool if possible
This should reduce the amount of TIME-WAIT sockets and TCP
handshakes when replicating, especially with small files.

An attempt was previously made to use the Net::HTTP::NB API
directly, but that resulted in complicated callback nesting
and state management needed to throttle the reader if the
sender socket were blocked in any way.

There were many bugs in the early version of this code as
a result of the complicated code.  Even after all the bugs
got fixed, a small performance reduction due to the extra
buffer copies was difficult to avoid.

Thus I started using the synchronous version to keep the code
simple and fast while preserving the ability to use persistent
sockets to avoid excessive TIME-WAIT and handshaking for small
file replication.