list_keys: consistent ESCAPE usage across DB types
Without specifying an ESCAPE character for LIKE queries, the '\'
we use for escaping is treated as a literal and improperly
matched keys with '\' in them under SQLite.

This is only needed for SQLite, as the SQLite language reference
makes no reference of a default ESCAPE character in
http://www.sqlite.org/lang_expr.html

ESCAPE is supported in MySQL and Postgres, too; and defaults to
'\'.  We specify it anyways to reduce code differences between
different databases.

Tested on MySQL 5.1.66 and Postgres 8.4.13 on Debian 6.0
and SQLite 3.7.13 on Debian 7.0