list_keys: escape in Store, allow [%\\] as prefix
If we support non-SQL DBs in the future, escaping rules could
become store-specific, so Worker/Query is not the right place for
it.

Since '%' and '\' may be escaped just like any other character,
we may also allow these characters as prefixes by properly
escaping them.

Tested on MySQL 5.1.66 and Postgres 8.4.13 on Debian 6.0
and SQLite 3.7.13 on Debian 7.0