ConnectionPool: improve reporting of socket creation errors
Send the entire error message (including intended host:port so
it is more informative when it propagates to
Connection::HTTP::err_response.  We also do not need to log
the error in ConnectionPool, as the error will be logged by
the caller.

While we're at it, fix the documentation and a spelling error in
err_response, too.