httpfile: correct timeouts for sidechannel digest
fsck digests are deprioritized and serialized in mogstored, so
it's nearly impossible to tell what's in the queue before our
request.  Since fsck is not latency critical, extend the timeout
for that.

We also need to account for normal seek/network latency for
non-fsck digest requests, so add node_timeout to that.

These bugs were mostly hidden since we are relying on <> to
read, which may incur watchdog timeouts.