store: do not auto-reconnect while holding a lock
Auto-reconnect is probably always unsafe while a holding a lock
on all networked databases.

While we do not use the builtin auto-reconnect functionality of
MySQL, any auto-reconnect implementation should be affected by
the same issues upon connection failure:
https://dev.mysql.com/doc/refman/5.6/en/auto-reconnect.html

With auto-reconnect, we could be operating under the false
assumption we have a lock after the reconnect when we do not.

For now, the easiest method of recovery is to just let the
worker die while working on the current task and have the
ProcManager restart it.