allow startup without job_master (and dependent workers)
And when running without a job_master, do not spawn job_master-dependent
workers (delete, fsck, replicate) as those workers will never get work.

Running a queryworker+monitor in a remote datacenter makes sense with
the MogileFS::Network plugins since the "create_close" size verification
is faster and more reliable if the queryworker is in the same datacenter
as the client, even if the master DB is in a remote datacenter.

Being in a remote datacenter, (master)DB-intensive operations from
delete, fsck and replicate workers can encounter high latency and an
unreliable link, so admins may disable those workers in this situation.

However, disabling delete, fsck, and replicate workers individually
still allows the job_master to fill the initial queues (which is never
processed) and prevent other trackers from processing items for 1000
seconds.

Future commits may allow job_master to ignore certain queues if there
are zero workers for that queue, but for now, stopping job_master
entirely should be sufficient for most users with trackers in a
different datacenter than the DB.

P.S. It also makes sense to disable reaper in remote datacenters, too,
but reaper does not rely on job_master.