add naive MultipleDevice replication policy
This can be useful when MultipleHosts is too noisy when hosts differ
greatly in storage capacity.

The intended target of this policy is a low-priority backup cluster
where a single host contains the bulk of the storage with a handful
of random machines helping out.  The MultipleHosts policy can be too
noisy with log messages about running out of suggestions in this case.