increase receive buffers for large state events
The monitor may send large state events for large installations
with many hosts, devices, domains, or classes.  The 1K default
is too small and leads to excessive syscalls and string operations.

This increases startup performance for a mock instance with 10K domains
and 10K non-default classes.

Using the parent_ping function and "No simple reply" warning as an
informal benchmark, this change reduces the loop time from 12 to 10
loops.