connection/poolable: do not write before event_write
Blindly attempting to write to a socket before a TCP connection can be
established returns EAGAIN on Linux, but not on FreeBSD 8/9.  This
causes Danga::Socket to error out, as it won't attempt to buffer on
anything but EAGAIN on write() attempts.

Now, we buffer writes explicitly after the initial socket creation and
connect(), and only call Danga::Socket::write when we've established
writability.  This works on Linux, too, and avoids an unnecessary
syscall in most cases.

Reported-by: Alex Yakovenko <aleksey.yakovenko@gmail.com>