connection/poolable: defer expiry of timed out connections
The timeout check may run on a socket before epoll_wait/kevent has
a chance to run, giving the application no chance for any readiness
callbacks to fire.

This prevents timeouts in the monitor if the database is slow during
synchronous UPDATE device calls (or there are just thousands of active
connections).