enable DB upgrade for host readonly state
The readonly host state was not enabled via mogdbsetup and
could not be used although the code supports it, making
the schema version bump to 16 a no-op.

This bumps the schema version to 17.

Add a test using mogadm to ensure the setting is changeable, as the
existing test for this state did not rely on the database.

This was also completely broken with Postgres before, as
Postgres currently offers no way to modify constraints in-place.
Constraints must be dropped and re-added instead.

Note: it seems the upgrade_add_device_* functions in Postgres.pm
are untested as well and never got used.  Perhaps they ought
to be removed entirely since those device columns predate Postgres
support.