client: use single write for admin commands
This avoids the odd case where the first write completes, but
the second one (for 3 bytes: ".\r\n") does not complete, causing
a client to having both read and write watchability enabled
after the previous commit to stop reads when writes do not
complete.

This would not be fatal, but breaks the rule where clients
should only be reading or writing exclusively, never doing
both; as that could lead to pathological memory usage.

This also reduces client wakeups and TCP overhead with
TCP_NODELAY sockets by avoiding a small packet (".\r\n")
after the main response.